builder/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
click==7.1.2              # via -r requirements.in
colorama==0.4.3           # via rich
commonmark==0.9.1         # via rich
pprintpp==0.4.0           # via rich
pycdlib==1.10.0           # via -r requirements.in
pygments==2.6.1           # via rich
rich==2.2.6               # via -r requirements.in
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.10      # via -r requirements.in
typing-extensions==3.7.4.2  # via rich
